feat: adds delivery publishing option
Adds a 3rd publishing option: delivery. The `deliver` function is
similar to `publish`, however it offers an explicit `ack` functiont to
acknowledge receipt, or completion. This not only allows for partial
processing (i.e. like HTTP Accepted), but also supports returning a
value, which can be useful in systems that have asynchronous message
sysetems for some features, and synchronous message systems for other.

As an example, Slack's command, event, and interactions features are all
asynchronous, but their call feature is synchronous, and requires a
slightly different response from Slack apps. Delivery is the best
publishing method for that.