fix: adds error reporting, resolves #7
Adds error reporting so that if a subscriber throws or rejects when
being emitted, published, or delivered, an "error" event is emitted.
This can be turned off by setting `reportVerbosity` to 'none' when
emitting something. It can be made more verbose by setting
`reportVerbosity` to 'all' when emitting something, which results in
fullfillment being emitted as well as errors.